Flask/requirements.txt
Flask==3.0.0
Flask_Cors==4.0.1
mysql_connector==2.2.9
Pillow==10.1.0
Werkzeug==2.2.2
